+++ src/fetchers/mod.rs
@@ -2,6 +2,7 @@
 pub mod events;
 pub mod games;
 pub mod historical_reports;
+pub mod official_stats;
 pub mod players;
 pub mod seasons;
 pub mod teams;